requirements.txt
google.generativeai
apikey
google-search-results
yahooquery
streamlit
yfinance
langchain-google-genai
langchain
langchain_community
